New Classes created and old classes updated
In this commit I have created new classes:
- Person
- Manager
- Position
- Employee
- Department
- DepartmentEmployee
- DepartmentPatient
- Salary

Employee:
Employee, extends Person, using the same fisrtname and lastname, removing the redundancy. In this I have created methods to get and set the Managertype, hours worked by the employee and calculate the salary.

Manager:
In this class i create a method to put in a List the department  in which each employee works, ex: Director Employee works in H.R department.
It will be useful at me moment when the user needs to insert a new employee and then he can pick one position and link it to a department.

Position:
Created to named the title of each position and create a list with allowedDepartments.

Salary:
Class created to set the amount of € per hour each managerType will earn.
I used the method calculateMonthlyPayment and a switch case to determine how much each position will receive per hour.
another method created to set up the amount of hours per month, more than 80 up to 160h/month.

Department, departmentEmployee and DepartmentPatient classes created to determine the departments where the employees will works and where the patient will be served.
Department is the Parent class and DepartmentPatient extends Department.
departmentEmployee extends departmentPatient, creating a relation between the area where the employee will works, in case there is a doctor(department Employee0 who will works in DepartmentPatient(Surgery) as exemple. So I decided to use this Inherance to link one department with another.