Create GenerateRandomEmployee.java
This classe was created to generate a random employee when the user select this option when using the menu for Employees.
This class will be able to generate a random name and surname collected from a .txt file that will be read when the user start the system.

Afterwards, this class would be able to generate a position and department randomly, creating all the informations abou this new employee to be added on the list at the end by the user.