Create Searcher.java
Class Searcher created and commented.

This class will be responsible to found the person on the list.
when the user enters the name and lastname, the binary search algorithm will check the list, dividing it in the middle every loop, until find the middle (person).

I created another method (searchByName) that will be used in the main class, and will receive the user String input and will use the binarySearchByName method to find the person.