Create Sorter.java
Class created to sort the list of employees in ascending order.
In this class, I decided to use the Mergesort algorithm, due to its efficiency and thinking about the future, when the database is larger than the current time, this algorithm can work well due to its O(n log n) time complexity, which it is really good in performance with a large volume of data.