Create Patient.java
Patient class created with the purpose of adding all the details of the care provided to the patient and having the entire record of the patient's admission to the hospital, including time, name, health condition, diagnosis, etc.

This information will then be inserted into the list, so that it can have the patient record and finally have the possibility of generating a new text file with the reports of all the patients treated, allowing future research to be carried out.