Methods Implemented in the main class and update in other classes
In this commit I have implemented all the methods that will be used in Menus in the main class.

And to complete and make sure all the functionalities in this program are working well, I update some classes as DataFile adding a new feature to write a report for patients and employees; create a new class only to run the program *ExecuteTheProgram" to be more clean when using it.
Update the generateRandomEmployee class adding variables like employeeName and employeeSurname to be more accurate when used in the main class.
Updated the searcher and sorter classes, fixing some issues when testing also.